requirements.txt
blinker==1.4
cachelib==0.1.1
certifi==2020.12.5
cffi==1.14.0
chardet==4.0.0
Click==7.0
Flask==1.1.1
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-sqlalchemy==0.24.1
mysql-connector-python==8.0.23
protobuf==3.14.0
pycparser==2.19
PyJWT==1.7.1
PyMySQL==1.0.2
pyodbc==4.0.30
requests==2.25.1
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.26.3
Werkzeug==1.0.0
